MAJOR REFACTOR
-Pulled out all the behaviour from the player script. it is now empty, but it has a designated new purpose, so it will not be removed
-Renamed the Jump component to Movement and put the walking code into it
-Created a component to handle eating/puking
-Created a component to handle restarting the level
