+++ PukingPredator/Assets/Scripts/Eating.cs
@@ -1,0 +1,88 @@
+using UnityEngine;
+
+[RequireComponent(typeof(Rigidbody))]
+public class Eating : InputBehaviour
+{
+    /// <summary>
+    /// The mass of the player before having eaten anything.
+    /// </summary>
+    private float baseMass;
+
+    /// <summary>
+    /// The layers that consumable objects can be on.
+    /// </summary>
+    [SerializeField]
+    private LayerMask consumableLayers;
+
+    /// <summary>
+    /// The players inventory.
+    /// </summary>
+    [SerializeField]
+    private Inventory inventory;
+
+    /// <summary>
+    /// The distance that items spawn ahead of the player when puking. (TEMP)
+    /// </summary>
+    private float pukeDistance = 2f;
+
+    /// <summary>
+    /// The rigidbody of the player.
+    /// </summary>
+    private Rigidbody rb;
+
+
+
+    void Start()
+    {
+        rb = GetComponent<Rigidbody>();
+
+        baseMass = rb.mass;
+
+        inventory.onChange += UpdateMass;
+
+        Subscribe(EventType.onEat, GameInput_Eat);
+        Subscribe(EventType.onPuke, GameInput_Puke);
+
+        Unsubscribe(EventType.onEat, GameInput_Eat);
+    }
+
+
+
+    private void GameInput_Eat()
+    {
+        if (inventory.isFull) { return; }
+
+        // Define the ray, starting from the player's position, shooting forward
+        var ray = new Ray(transform.position, transform.forward);
+        RaycastHit hit;
+        if (Physics.SphereCast(ray, radius: 0.5f, out hit, maxDistance: 5f, layerMask: consumableLayers))
+        {
+            // Get the GameObject that was hit
+            GameObject hitObject = hit.collider.gameObject;
+
+            var consumableData = hitObject.GetComponent<Consumable>();
+            if (consumableData == null || !consumableData.isConsumable) { return; }
+
+            inventory.PushItem(consumableData);
+            consumableData.SetState(ItemState.beingConsumed);
+        }
+    }
+
+    private void GameInput_Puke()
+    {
+        if (inventory.isEmpty) { return; }
+
+        Consumable itemToPlace = inventory.PopItem();
+
+        var pukeDir = transform.forward;
+        itemToPlace.PlaceAt(transform.position + pukeDir * pukeDistance);
+    }
+
+    private void UpdateMass()
+    {
+        var totalItemMass = inventory.totalMass;
+
+        //TODO: make this update the mass, doesnt have to be 1-1 or even linear
+        rb.mass = baseMass;
+    }
+}
