+++ PukingPredator/Assets/Scripts/Eating.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 0c7f26f8eaf9849399a218ec7cfb8b6e
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
