+++ PukingPredator/Assets/Scripts/Movement.cs
@@ -1,7 +1,7 @@
 using UnityEngine;
 
 [RequireComponent(typeof(Rigidbody))]
-public class Jump : MonoBehaviour
+public class Movement : InputBehaviour
 {
     /// <summary>
     /// The max time to hold the jump button for.
@@ -13,12 +13,6 @@ public class Jump : MonoBehaviour
     /// If the instance can currently jump.
     /// </summary>
     private bool canJump => isGrounded && !isJumping;
-
-    /// <summary>
-    /// Force applied downwards to reduce jump height if you let go early.
-    /// </summary>
-    [SerializeField]
-    private float cancelRate = 25;
 
     /// <summary>
     /// Radius of the sphere used for collision checks.
@@ -47,6 +41,12 @@ public class Jump : MonoBehaviour
     private bool isJumping;
 
     /// <summary>
+    /// Force applied downwards to reduce jump height if you let go early.
+    /// </summary>
+    [SerializeField]
+    private float jumpCancelRate = 25;
+
+    /// <summary>
     /// The force applied when jumping.
     /// </summary>
     [SerializeField]
@@ -58,6 +58,22 @@ public class Jump : MonoBehaviour
     private float jumpTime;
 
     /// <summary>
+    /// The direction of the players recent movement input with the camera rotation applied.
+    /// </summary>
+    private Vector3 moveDir;
+
+    /// <summary>
+    /// The speed that the player moves in any given direction.
+    /// </summary>
+    private float moveSpeed = 10;
+
+    /// <summary>
+    /// A reference to the camera, used for correcting movement direction.
+    /// </summary>
+    [SerializeField]
+    private GameObject playerCamera;
+
+    /// <summary>
     /// The rigidbody of the instance.
     /// </summary>
     private Rigidbody rb;
@@ -67,10 +83,37 @@ public class Jump : MonoBehaviour
     private void Start()
     {
         rb = GetComponent<Rigidbody>();
+
+        if (playerCamera == null) { playerCamera = GameObject.FindGameObjectsWithTag("MainCamera")[0]; }
+    }
+
+    private void FixedUpdate()
+    {
+        //walking code
+        //TODO: make this depend on the mass of the object
+        rb.velocity = moveDir * moveSpeed + new Vector3(0, rb.velocity.y, 0);
+
+        //jumping code
+        //Reduce jump height if the button is released early
+        if (isJumpCancelled && isJumping && rb.velocity.y > 0)
+        {
+            rb.AddForce(Vector3.down * jumpCancelRate);
+            if (rb.velocity.y <= 0) { isJumping = false; }
+        }
     }
 
     private void Update()
     {
+        if (gameInput == null) { return; }
+
+        //walking code
+        Vector2 inputVector = gameInput.movementInput;
+        moveDir = Quaternion.Euler(0, playerCamera.transform.eulerAngles.y, 0) * new Vector3(inputVector.x, 0, inputVector.y);
+
+        float turnSpeed = 10f;
+        transform.forward = Vector3.Slerp(transform.forward, moveDir, Time.deltaTime * turnSpeed);
+
+        //jumping code
         isGrounded = Physics.CheckSphere(transform.position, groundCheckRadius, groundLayer);
 
         if (Input.GetKeyDown(KeyCode.Space) && canJump)
@@ -90,15 +133,4 @@ public class Jump : MonoBehaviour
             if (jumpTime > buttonTime) { isJumping = false; }
         }
     }
-
-    private void FixedUpdate()
-    {
-        //Reduce jump height if the button is released early
-        if (isJumpCancelled && isJumping && rb.velocity.y > 0)
-        {
-            rb.AddForce(Vector3.down * cancelRate);
-            if (rb.velocity.y <= 0) { isJumping = false; }
-        }
-    }
-
 }
