+++ PukingPredator/Assets/Scripts/RestartLevel.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: f71f7d57b8c3c44f78f4a26b23851fe2
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
